Added kata 06 anagrams, working towards improved memory management
Using Valgrind.

+++ src/file/read_file.h
@@ -2,8 +2,8 @@
 #define KATA_READ_FILE_H
 
 typedef struct strings_list {
-    int lines_count;
-    int max_lines;
+    size_t lines_count;
+    size_t max_lines;
     char **lines;
 } StringsList;
